server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
python-multipart==0.0.6
PyMuPDF==1.23.8
python-docx==1.1.0
pdfminer.six==20231228
pytesseract==0.3.10
Pillow==10.1.0
botocore==1.34.162
aiofiles==23.2.0
requests==2.31.0
python-dotenv==1.0.0